pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jackbutler"
version = "0.1.0"
description = "Guitar Pro Music Theory Analyzer"
requires-python = ">=3.11"
dependencies = [
    "pyguitarpro>=0.9",
    "music21>=9.1",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.18",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/jackbutler"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
